Update vulnerable plugin dependency using constraint
The 'com.github.ben-manes.versions' plugin' has a dependency on a vulnerable version
of 'com.squareup.okio:okio'.

Assuming there is no fixed version of the plugin available,
we avoid this vulnerability using a dependency constraint that causes
a newer version of 'okio' to be used in the buildscript classpath.