Share dependency-submission workflow
Previously the dependency-submission action was used in 2 different workflows,
one for dependency-submission on pushes to `main` and another for depenency-review.
This setup can lead to duplicate dependencies in the GitHub Dependency Insights,
due to some odd behaviour in handling dependency submission.

We now use a single workflow Job for all dependency-submission, and use the
dependency-review-action in the "wait for snapshot" mode.

This setup is now recommended in the docs for `gradle/actions/dependency-submission`.